Multiple changes. Finished NGDebug/Assert library, refactored, added editor window for managing config file, finished settings save/load. Refactored ShowLocalAxis and its inspector script.
