+++ .gitignore
@@ -20,3 +20,5 @@
 # Unity3D Generated File On Crash Reports
 sysinfo.txt
 NGTools_Unity/NGTools/Assets/UnityVS/
+NGTools_Unity/NGTools/Assets/UnityTestTools/
+NGTools_Unity/NGTools/logs/
