+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/Editor/NGDebug.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: ddff2a75ff25d2242bd882d81710a92d
+folderAsset: yes
+timeCreated: 1446625402
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
