+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/Editor/NGDebug/NGDebugEditorWindow.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a5b28273b420cb64a9f41648995c4650
+timeCreated: 1446625693
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
