+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/NGDebug/DebugConsole.cs
