+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/NGDebug/NGAssertDebugConfig.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 8a3b01b0c39a5934188de776b8a70050
+timeCreated: 1446625402
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
