+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/NGDebug/NGDebug.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d3deea8d21d5c7848875f0b0d8f5fa21
-timeCreated: 1445847138
+guid: a53a48da5f61f6b4a986a8d64d52ba4c
+timeCreated: 1446625402
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
