+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/NGDebug/NGDebugToFile.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 2d646998d0759714fa2158a9d647bd50
+timeCreated: 1446625402
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
