+++ NGTools_Unity/NGTools/Assets/_Scripts/NGTools/NGDebug/NGFailureHandler.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a0628e12c2b3ea444bf7e887bc5b34a3
+timeCreated: 1446625402
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
